+++ rasa_core/events/__init__.py
@@ -15,6 +15,7 @@ import jsonpickle
 from dateutil import parser
 
 from rasa_core import utils
+from rasa_nlu.training_data.formats import MarkdownWriter, MarkdownReader
 
 if typing.TYPE_CHECKING:
     from rasa_core.trackers import DialogueStateTracker
@@ -42,6 +43,23 @@ def deserialise_events(serialized_events):
                                "events. Couldn't parse it.")
 
     return deserialised
+
+
+def deserialise_entities(entities):
+    if isinstance(entities, str):
+        entities = json.loads(entities)
+
+    return [e for e in entities if isinstance(e, dict)]
+
+
+def md_format_message(text, intent, entities):
+    message_from_md = MarkdownReader()._parse_training_example(text)
+    deserialised_entities = deserialise_entities(entities)
+    return MarkdownWriter()._generate_message_md(
+            {"text": message_from_md.text,
+             "intent": intent,
+             "entities": deserialised_entities}
+    )
 
 
 def first_key(d, default_key):
@@ -230,7 +248,7 @@ class UserUttered(Event):
         except KeyError as e:
             raise ValueError("Failed to parse bot uttered event. {}".format(e))
 
-    def as_story_string(self):
+    def as_story_string(self, e2e=False):
         if self.intent:
             if self.entities:
                 ent_string = json.dumps({ent['entity']: ent['value']
@@ -238,9 +256,16 @@ class UserUttered(Event):
             else:
                 ent_string = ""
 
-            return "{intent}{entities}".format(
-                    intent=self.intent.get("name", ""),
-                    entities=ent_string)
+            parse_string = "{intent}{entities}".format(
+                           intent=self.intent.get("name", ""),
+                           entities=ent_string)
+            if e2e:
+                message = md_format_message(self.text,
+                                            self.intent,
+                                            self.entities)
+                return "{}: {}".format(self.intent.get("name"), message)
+            else:
+                return parse_string
         else:
             return self.text